auditing the double messaging issue 2
